backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
pymongo==4.15.3
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==4.0.1
pydantic-settings==2.11.0
email-validator==2.3.0
pytest==8.4.2
httpx==0.28.1
mongomock==4.3.0
google-auth==2.41.1
requests==2.32.3

